Introduce big number adder to Fibonacci driver
In order to handle requests that ask for Fibonacci (93) or greater,
a big number adder is introduced to this kernel module.
This big number adder is based on xs string implemented on this
repo: https://github.com/AdrianHuang/fibdrv.git. But in this
implementation, a byte is used to store 2 digits rather than 1
digit, which will half the space consumed for a single number.
Function bn_new aims to initilize the given big number with a C
string styled number. The core function of big number is bn_add,
which performs addition between two big numbers. To withdraw the
content of big number, a character buffer is required, whose size
could be determined by the return value of bn_space_for_buf.
After all, ones can withdraw the string content by invoking
bn_to_digit_string with buffer built before.
In order to return the correct value back to the user, the result
is now returned by writing to user buffer, rather than by return
value.